requirements.txt
xarray
netCDF4
numpy
pandas
pyarrow
sqlalchemy
psycopg2-binary
geoalchemy2
python-dateutil
loguru
s3fs
python-dotenv
matplotlib
